pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ciphersweep"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.9.0",
  "tqdm>=4.66.0",
  "cryptography>=42.0.0",
  "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["ciphersweep*"]

[project.scripts]
ciphersweep = "ciphersweep.cli:app"
